Module Sihl_facade__Migration

val to_sexp : Sihl_contract.Migration.t -> Sexplib0.Sexp.t

val pp : Stdlib.Format.formatter -> Sihl_contract.Migration.t -> unit

val empty : string -> Sihl_contract.Migration.t

val create_step : label:string -> ?⁠check_fk:bool -> string -> Sihl_contract.Migration.step

val add_step : Sihl_contract.Migration.step -> Sihl_contract.Migration.t -> Sihl_contract.Migration.t

include Sihl_contract.Migration.Sig

include Sihl_core.Container.Service.Sig

val lifecycle : Sihl_core.Container.Lifecycle.t

val register_migration : Sihl_contract.Migration.t -> unit

register_migration migration registers a migration migration with the migration service so it can be executed with `run_all`.

val register_migrations : Sihl_contract.Migration.t list -> unit

register_migrations migrations registers migrations migrations with the migration service so it can be executed with `run_all`.

val execute : Sihl_contract.Migration.t list -> unit Lwt.t

execute migrations runs all migrations migrations.

val run_all : unit -> unit Lwt.t

run_all () runs all migrations that have been registered.

val register : ?⁠migrations:Sihl_contract.Migration.t list -> unit -> Sihl_core.Container.Service.t

val lifecycle : unit -> Sihl_core.Container.Lifecycle.t

val register : ?⁠migrations:Sihl_contract.Migration.t list -> (module Sihl_contract.Migration.Sig) -> Sihl_core.Container.Service.t